+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 **Added**
 
+- Introduced base models for persons, transactions, incomes, and expenses,
+  with transaction categories and a transaction value computing method.
 - Introduced an enhanced logger with progress tracking with methods for logging messages and warnings,
   setting/removing log prefixes, and tracking progress of data iteration with a progress bar.
 - Implemented a slugify utility to generate identifiers.